updated JUCE source and Visual Studio project files, also fixed plugin bugs on windows

+++ JuceLibraryCode/modules/juce_gui_basics/windows/juce_TooltipWindow.cpp
@@ -183,7 +183,7 @@ void TooltipWindow::timerCallback()
         // appear after a timeout..
         if (newTip.isNotEmpty()
              && newTip != tipShowing
-             && now > lastCompChangeTime + (unsigned int) millisecondsBeforeTipAppears)
+             && now > lastCompChangeTime + millisecondsBeforeTipAppears)
         {
             showFor (newTip);
         }
